feat: cmd scan
